Directory breadcrumbs
Warning: This only works when browsing a master branch!  I couldn't find a
clean way to pass three arguments into a filter, so I left out the changeset
and defaulted it to master.  I'm considering changing the path_breadcrumbs
filter into a template tag.  That would clean things up a lot and make it
easier to pass arguments, but the implementation would be more complicated.

Filing an issue to fix this.